requirements.txt
discord.py
yt-dlp
aiohttp
beautifulsoup4
pillow
requests